# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/units": {
      "get": {
        "operationId": "Units_list",
        "summary": "Units in the caller's scope",
        "description": "A04, DL04, CU02. `q` matches serial, customer, dealer and model code. Default sort `serial`.",
        "parameters": [
          {
            "name": "dealerId",
            "required": false,
            "in": "query",
            "schema": {}
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "ACTIVE",
                "EXPIRING_SOON",
                "EXPIRED",
                "VOID",
                "PENDING"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated<UnitView>"
          }
        },
        "tags": [
          "units"
        ]
      }
    },
    "/api/units/{serial}": {
      "get": {
        "operationId": "Units_get",
        "summary": "Unit with part-wise warranty status for today",
        "description": "404 if not visible.",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "UnitView"
          }
        },
        "tags": [
          "units"
        ]
      }
    },
    "/api/units/{serial}/certificate.pdf": {
      "get": {
        "operationId": "Units_certificate",
        "summary": "Warranty certificate PDF",
        "description": "Bearer token or the session cookie.",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "application/pdf"
          }
        },
        "tags": [
          "units"
        ]
      }
    },
    "/api/units/{serial}/entitlement": {
      "get": {
        "operationId": "Units_entitlement",
        "summary": "What a complaint on this unit would be entitled to today (CU04, DL06)",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Entitlement"
          }
        },
        "tags": [
          "units"
        ]
      }
    },
    "/api/units/{serial}/void": {
      "post": {
        "operationId": "Units_voidWarranty",
        "summary": "Void the warranty (W5)",
        "description": "`422` validation.voidReason; `409 already_void`; `409 not_registered`.",
        "parameters": [
          {
            "name": "serial",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "reason"
                ],
                "properties": {
                  "reason": {
                    "type": "string",
                    "enum": [
                      "UNAUTHORISED_REPAIR",
                      "MISSED_SERVICING",
                      "PHYSICAL_DAMAGE",
                      "OTHER"
                    ]
                  },
                  "note": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "UnitView with `void` set"
          }
        },
        "tags": [
          "units"
        ]
      }
    },
...
  },
  "info": {
    "title": "Warranty Management API",
    "description": "Units with part-wise warranties, registrations from every channel, complaints, service hand-off, manufacturer claims and the integration log. Contract: frontend/docs/api-contract.md.",
    "version": "0.2.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "wms_refresh"
      }
    },
...
  }
}